requirements.txt
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
idna==3.3
isodate==0.6.0
msrest==0.6.21
numpy==1.21.4
oauthlib==3.1.1
opencv-python-headless==4.5.4.60
pycparser==2.21
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
